# Optimize encoding functions for unescaped strings